EON-1981 fix issues with rebuilding indexes
Technically 2 fixes here 1 bug and 1 work around.

The work around (in create_actions) deals with a non-critical error being
propigated by elasticsearch.py that I believe is fixed in later builds but
is outside the scope of this task.

The bug fix deals with ES 2x treats attempting to alter merge settings
as an error where as 1x did not (although it also did not change them).